# api_wrapper/openapi.yaml
openapi: "3.0.3"
info:
  title: "A-Stock Data API — 24 endpoints, 7 layers"
  version: "3.3.0"
  description: >
    A 股全栈数据 API，专为大模型 Function Calling / Tool Use 设计。
    覆盖行情、研报、信号、资金面、新闻、基础数据、公告七大层级。

    数据源优先级：腾讯/百度/同花顺（不封IP）优先，东财仅用于独有数据（内置 em_get 串行节流 ≥1s）。
    全路由容错降级：异常时返回 HTTP 200 + 空结构 + note，绝不抛 500。

servers:
  - url: http://8.163.120.79:9000
    description: 阿里云 ECS 生产环境

tags:
  - name: Quote
    description: 实时行情 — 腾讯/百度（永不被封IP）
  - name: Signal
    description: 信号层 — 热点归因/北向资金/概念/龙虎榜/行业
  - name: Fund
    description: 资金面 — 融资融券/大宗交易/解禁/股东户数/分红/资金流向
  - name: Research
    description: 研报层 — 研报列表/EPS一致预期/iwencai语义搜索
  - name: News
    description: 新闻层 — 个股新闻/全球7x24资讯
  - name: Fundamentals
    description: 基础数据 — 财报三表/公告/估值
  - name: System
    description: 系统 — 健康检查/防封状态

paths:
  # ============================
  # Quote Layer
  # ============================
  /api/v1/stock/quote:
    get:
      tags: [Quote]
      operationId: getQuote
      summary: 腾讯财经实时行情（永不被封IP）
      description: |
        批量拉取 PE(TTM)/PB/市值/换手率/涨跌停价。
        支持个股、指数（000001上证/000300沪深300/399006创业板）、ETF。
      parameters:
        - name: codes
          in: query
          required: true
          description: 股票代码，逗号分隔。如 688017,600519,000858
          schema:
            type: string
            example: "688017,600519"

  /api/v1/stock/kline:
    get:
      tags: [Quote]
      operationId: getKline
      summary: 百度K线 + MA5/MA10/MA20 均价
      description: 百度股市通日K线，自带移动均线，无需本地计算。
      parameters:
        - name: code
          in: query
          required: true
          description: 6位股票代码
          schema:
            type: string
            pattern: "^[0-9]{6}$"
            example: "600519"
        - name: start_time
          in: query
          required: false
          description: 起始日期 YYYY-MM-DD，空=全部
          schema:
            type: string
            example: "2026-01-01"

  # ============================
  # Signal Layer
  # ============================
  /api/v1/stock/hot:
    get:
      tags: [Signal]
      operationId: getHotStocks
      summary: 当日强势股归因（同花顺，零鉴权73ms）
      description: |
        返回当日所有强势股 + 题材标签 reason。
        约125只股票，含涨幅/换手/大单净量等字段。
      parameters:
        - name: date
          in: query
          required: false
          description: 日期 YYYY-MM-DD，默认今天
          schema:
            type: string
            example: "2026-06-03"

  /api/v1/stock/northbound:
    get:
      tags: [Signal]
      operationId: getNorthbound
      summary: 北向资金实时分钟流向
      description: |
        沪深股通当日累计净买入（亿元），262个时间点（09:10-15:00）。
        返回 summary.hgt_yi（沪股通）、sgt_yi（深股通）、total_yi（合计）。

  /api/v1/stock/concept:
    get:
      tags: [Signal]
      operationId: getConcept
      summary: 概念板块归属（东财 slist，V3.2.2）
      description: |
        个股所属全部板块（行业/概念/地域混合），一次请求拿全。
        返回 total、boards（含 name/BK码/涨跌幅/龙头股）、concept_tags。
        接口受限时降级为仅行业信息，再失败返回空盒（不抛500）。
      parameters:
        - name: code
          in: query
          required: true
          description: 6位股票代码
          schema:
            type: string
            pattern: "^[0-9]{6}$"
            example: "688017"

  /api/v1/stock/industry:
    get:
      tags: [Signal]
      operationId: getIndustryRanking
      summary: 行业板块涨跌排名（东财push2）
      description: 全行业涨跌幅排名，含领涨股。top=涨幅前N，bottom=跌幅前N。
      parameters:
        - name: top_n
          in: query
          required: false
          description: 取前/后N名，默认20
          schema:
            type: integer
            default: 20

  /api/v1/stock/dragon_tiger:
    get:
      tags: [Signal]
      operationId: getDragonTiger
      summary: 龙虎榜（个股席位 + 全市场）
      description: |
        scope=single: 个股近N日上榜记录 + 买卖席位TOP5 + 机构汇总
        scope=all: 全市场龙虎榜，按净买入降序
      parameters:
        - name: code
          in: query
          required: false
          description: 6位代码。scope=single时必填
          schema:
            type: string
            example: "002475"
        - name: scope
          in: query
          required: false
          description: single(个股) / all(全市场)
          schema:
            type: string
            enum: [single, all]
            default: single
        - name: trade_date
          in: query
          required: false
          description: 日期 YYYY-MM-DD，默认今天
          schema:
            type: string
        - name: look_back
          in: query
          required: false
          description: 回看天数(scope=single)，默认30
          schema:
            type: integer
            default: 30
        - name: min_net_buy
          in: query
          required: false
          description: 净买入下限万元(scope=all)
          schema:
            type: number

  # ============================
  # Fund Layer
  # ============================
  /api/v1/stock/fund_flow:
    get:
      tags: [Fund]
      operationId: getFundFlow
      summary: 个股资金流向（分钟级 + 120日）
      description: |
        period=minute: 当日逐分钟主力/超大单/大单/中单/小单净流入（元）
        period=120d: 近120交易日每日汇总
      parameters:
        - name: code
          in: query
          required: true
          description: 6位股票代码
          schema:
            type: string
            example: "000858"
        - name: period
          in: query
          required: false
          description: minute(分钟级) / 120d(日级)
          schema:
            type: string
            enum: [minute, "120d"]
            default: minute

  /api/v1/stock/margin:
    get:
      tags: [Fund]
      operationId: getMarginTrading
      summary: 融资融券明细
      description: 融资余额/融资买入/融券余额（日级，元）。rzye持续增加=看多信号。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "600519"
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 30

  /api/v1/stock/block_trade:
    get:
      tags: [Fund]
      operationId: getBlockTrade
      summary: 大宗交易记录
      description: 含成交价/收盘价/溢价率/买卖方名称。premium_pct>0=溢价成交（利好信号）。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "600519"
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 20

  /api/v1/stock/lockup:
    get:
      tags: [Fund]
      operationId: getLockupExpiry
      summary: 限售解禁日历
      description: history=历史解禁，upcoming=未来N天待解禁。大额解禁=潜在抛压。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "002475"
        - name: trade_date
          in: query
          required: false
          description: 基准日期，默认今天
          schema:
            type: string
        - name: forward_days
          in: query
          required: false
          description: 向前看天数，默认90
          schema:
            type: integer
            default: 90

  /api/v1/stock/holder:
    get:
      tags: [Fund]
      operationId: getHolderChange
      summary: 股东户数变化
      description: change_ratio为负=筹码集中（主力吸筹信号），为正=筹码分散。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "600519"
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 20

  /api/v1/stock/dividend:
    get:
      tags: [Fund]
      operationId: getDividendHistory
      summary: 分红送转历史
      description: bonus_rmb=每股派息（税前），transfer_ratio=每10股转增，bonus_ratio=每10股送股。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "600519"
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 20

  # ============================
  # Research Layer
  # ============================
  /api/v1/stock/reports:
    get:
      tags: [Research]
      operationId: getReports
      summary: 研报列表（东财reportapi）
      description: 拉取指定股票所有研报，含机构名称/评级/发布日期/title。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "688017"
        - name: max_pages
          in: query
          required: false
          description: 最大翻页数，默认5
          schema:
            type: integer
            default: 5

  /api/v1/stock/eps_forecast:
    get:
      tags: [Research]
      operationId: getEpsForecast
      summary: 机构一致预期EPS（同花顺）
      description: |
        返回未来年份的EPS预测：analysts(机构数)/min_eps/mean_eps/max_eps。
        analysts<3=覆盖不足，需谨慎。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "688017"

  /api/v1/stock/iwencai/search:
    get:
      tags: [Research]
      operationId: iwencaiSearch
      summary: iwencai语义搜索（需API Key）
      description: |
        NL语义搜索研报/公告/新闻。
        channel: report(研报)/announcement(公告)/news(新闻)
        需要环境变量 IWENCAI_API_KEY，申请: https://www.iwencai.com/skillhub
      parameters:
        - name: query
          in: query
          required: true
          description: 自然语言查询，如 "人形机器人 行星滚柱丝杠 2026"
          schema:
            type: string
        - name: channel
          in: query
          required: false
          schema:
            type: string
            enum: [report, announcement, news]
            default: report
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50

  /api/v1/stock/iwencai/query:
    get:
      tags: [Research]
      operationId: iwencaiQuery
      summary: iwencai结构化数据查询（需API Key）
      description: |
        NL查询返回结构化数据行，如 "贵州茅台 ROE" → 表格数据。
        需要环境变量 IWENCAI_API_KEY。
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            example: "贵州茅台 ROE"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50

  # ============================
  # News Layer
  # ============================
  /api/v1/stock/news:
    get:
      tags: [News]
      operationId: getStockNews
      summary: 个股新闻
      description: 东财个股相关新闻，含标题/摘要/来源/时间。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "688017"
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 20

  /api/v1/stock/global_news:
    get:
      tags: [News]
      operationId: getGlobalNews
      summary: 全球财经资讯（7x24滚动）
      description: 东财全球快讯，实时滚动更新。
      parameters:
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 50

  # ============================
  # Fundamentals Layer
  # ============================
  /api/v1/stock/financial:
    get:
      tags: [Fundamentals]
      operationId: getFinancialReport
      summary: 财报三表（新浪）
      description: |
        type=lrb(利润表)/fzb(资产负债表)/llb(现金流量表)
        返回最近N期数据，含同比。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "600519"
        - name: type
          in: query
          required: false
          description: lrb(利润表)/fzb(资产负债表)/llb(现金流量表)
          schema:
            type: string
            enum: [lrb, fzb, llb]
            default: lrb
        - name: num
          in: query
          required: false
          description: 取最近N期，默认8
          schema:
            type: integer
            default: 8

  /api/v1/stock/announcements:
    get:
      tags: [Fundamentals]
      operationId: getAnnouncements
      summary: 公告全文检索（巨潮）
      description: 沪/深/北交所公告，含标题/类型/日期/链接。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "600519"
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            default: 30

  /api/v1/stock/valuation:
    get:
      tags: [Fundamentals]
      operationId: getValuation
      summary: 完整估值分析（PE/PEG/消化年数）
      description: |
        一站式返回：最新价/市值/PE(TTM)/PB/一致预期EPS/前向PE/CAGR/PEG/消化年数。
        需pandas（ECS已装）。
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: "688017"

  /api/v1/stock/peg:
    get:
      tags: [Fundamentals]
      operationId: calcPeg
      summary: PEG 计算器
      description: |
        传入PE和CAGR(小数，如0.15=15%)，返回PEG和消化年数。
        PEG<1便宜，1-1.5合理，>1.5贵。
      parameters:
        - name: pe
          in: query
          required: true
          schema:
            type: number
            example: 25
        - name: cagr
          in: query
          required: true
          description: 增长率（小数），如0.15=15%
          schema:
            type: number
            example: 0.15

  # ============================
  # System
  # ============================
  /api/v1/health:
    get:
      tags: [System]
      operationId: healthCheck
      summary: 健康检查 + 防封状态
      description: 返回服务运行状态、东财API累计调用次数、调用频率、运行时长。